Implementando la edicion desde el front, testeando con la seccion Experiencia laboral

+++ BackEnd/src/main/java/com/Portfolio/DavKol/Repository/IPersonaRepository.java
@@ -1,4 +1,3 @@
-
 package com.Portfolio.DavKol.Repository;
 
 import com.Portfolio.DavKol.Entity.Persona;
@@ -6,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPersonaRepository extends JpaRepository <Persona,Long> {
-    
+public interface IPersonaRepository extends JpaRepository <Persona,Integer> {    
 }
